Switch iBytes argument from uint64 to int64
io.Copy returns int64, so the n values fed to iBytes lose their
explicit casts. The remaining float64/int call sites still need a
cast, but int64 lines up with the rest of the test code.

+++ shapeio_test.go
@@ -17,12 +17,12 @@ import (
 // iBytes formats b as an IEC binary size string (e.g. "1.0 MiB"). It is a
 // tiny replacement for humanize.IBytes, kept here so the test suite has no
 // external dependency.
-func iBytes(b uint64) string {
+func iBytes(b int64) string {
 	const unit = 1024
 	if b < unit {
 		return fmt.Sprintf("%d B", b)
 	}
-	div, exp := uint64(unit), 0
+	div, exp := int64(unit), 0
 	for n := b / unit; n >= unit; n /= unit {
 		div *= unit
 		exp++
@@ -101,10 +101,10 @@ func TestRead(t *testing.T) {
 			}
 			t.Logf(
 				"read %s / %s: Real %s/sec Limit %s/sec. (%f %%)",
-				iBytes(uint64(n)),
+				iBytes(n),
 				elapsed,
-				iBytes(uint64(realRate)),
-				iBytes(uint64(limit)),
+				iBytes(int64(realRate)),
+				iBytes(int64(limit)),
 				realRate/limit*100,
 			)
 		}
@@ -143,7 +143,7 @@ func TestDynamicReadRateLimit(t *testing.T) {
 	if elapsed >= 5*time.Second {
 		t.Errorf("expected dynamic rate change to finish well under 5s, took %s", elapsed)
 	}
-	t.Logf("dynamic read: %s in %s", iBytes(uint64(n)), elapsed)
+	t.Logf("dynamic read: %s in %s", iBytes(n), elapsed)
 }
 
 // TestDynamicWriteRateLimit is the writer counterpart of TestDynamicReadRateLimit.
@@ -185,7 +185,7 @@ func TestDynamicWriteRateLimit(t *testing.T) {
 	if elapsed >= 5*time.Second {
 		t.Errorf("expected dynamic rate change to finish well under 5s, took %s", elapsed)
 	}
-	t.Logf("dynamic write: %s in %s", iBytes(uint64(written)), elapsed)
+	t.Logf("dynamic write: %s in %s", iBytes(int64(written)), elapsed)
 }
 
 // TestConcurrentSetRateLimitInitial exercises the case where SetRateLimit and
@@ -295,7 +295,7 @@ func TestSetRateLimitEvery(t *testing.T) {
 		if realRate > expected*(1+rateTolerance) || realRate < expected*(1-rateTolerance) {
 			t.Errorf("Limit %f but real rate %f (outside +/- %.0f%%)", expected, realRate, rateTolerance*100)
 		}
-		t.Logf("read %s in %s -> %s/sec", iBytes(uint64(n)), elapsed, iBytes(uint64(realRate)))
+		t.Logf("read %s in %s -> %s/sec", iBytes(n), elapsed, iBytes(int64(realRate)))
 	})
 
 	t.Run("Writer", func(t *testing.T) {
@@ -315,7 +315,7 @@ func TestSetRateLimitEvery(t *testing.T) {
 		if realRate > expected*(1+rateTolerance) || realRate < expected*(1-rateTolerance) {
 			t.Errorf("Limit %f but real rate %f (outside +/- %.0f%%)", expected, realRate, rateTolerance*100)
 		}
-		t.Logf("wrote %s in %s -> %s/sec", iBytes(uint64(n)), elapsed, iBytes(uint64(realRate)))
+		t.Logf("wrote %s in %s -> %s/sec", iBytes(n), elapsed, iBytes(int64(realRate)))
 	})
 }
 
@@ -343,10 +343,10 @@ func TestWrite(t *testing.T) {
 			}
 			t.Logf(
 				"write %s / %s: Real %s/sec Limit %s/sec. (%f %%)",
-				iBytes(uint64(n)),
+				iBytes(n),
 				elapsed,
-				iBytes(uint64(realRate)),
-				iBytes(uint64(limit)),
+				iBytes(int64(realRate)),
+				iBytes(int64(limit)),
 				realRate/limit*100,
 			)
 		}
